// packages/frontend/src/api/financialModelingPrep/types.ts
export type IncomeStatement = {
   date: string
   symbol: string
   reportedCurrency: string
   fillingDate: string
   acceptedDate: string
   period: string
   calendarYear?: string
   revenue: number
   costOfRevenue: number
   grossProfit: number
   grossProfitRatio: number
   researchAndDevelopmentExpenses: number
   generalAndAdministrativeExpenses: number
   sellingAndMarketingExpenses: number
   otherExpenses: number
   operatingExpenses: number
   costAndExpenses: number
   interestExpense: number
   depreciationAndAmortization: number
   ebitda: number
   ebitdaratio: number
   operatingIncome: number
   operatingIncomeRatio: number
   totalOtherIncomeExpensesNet: number
   incomeBeforeTax: number
   incomeBeforeTaxRatio: number
   incomeTaxExpense: number
   netIncome: number
   netIncomeRatio: number
   eps: number
   epsdiluted: number
   weightedAverageShsOut: number
   weightedAverageShsOutDil: number
   link: string
   finalLink: string
}

export type EnterpriseValue = {
   symbol: string
   date: string
   stockPrice: number
   numberOfShares: number
   marketCapitalization: number
   minusCashAndCashEquivalents: number
   addTotalDebt: number
   enterpriseValue: number
}

export type EnterpriseRatio = {
   symbol: string
   date: string
   period: string
   currentRatio: number
   quickRatio: number
   cashRatio: number
   daysOfSalesOutstanding: number
   daysOfInventoryOutstanding: number
   operatingCycle: number
   daysOfPayablesOutstanding: number
   cashConversionCycle: number
   grossProfitMargin: number
   operatingProfitMargin: number
   pretaxProfitMargin: number
   netProfitMargin: number
   effectiveTaxRate: number
   returnOnAssets: number
   returnOnEquity: number
   returnOnCapitalEmployed: number
   netIncomePerEBT: number
   ebtPerEbit: number
   ebitPerRevenue: number
   debtRatio: number
   debtEquityRatio: number
   longTermDebtToCapitalization: number
   totalDebtToCapitalization: number
   interestCoverage: number
   cashFlowToDebtRatio: number
   companyEquityMultiplier: number
   receivablesTurnover: number
   payablesTurnover: number
   inventoryTurnover: number
   fixedAssetTurnover: number
   assetTurnover: number
   operatingCashFlowPerShare: number
   freeCashFlowPerShare: number
   cashPerShare: number
   payoutRatio: number
   operatingCashFlowSalesRatio: number
   freeCashFlowOperatingCashFlowRatio: number
   cashFlowCoverageRatios: number
   shortTermCoverageRatios: number
   capitalExpenditureCoverageRatio: number
   dividendPaidAndCapexCoverageRatio: number
   dividendPayoutRatio: number
   priceBookValueRatio: number
   priceToBookRatio: number
   priceToSalesRatio: number
   priceEarningsRatio: number
   priceToFreeCashFlowsRatio: number
   priceToOperatingCashFlowsRatio: number
   priceCashFlowRatio: number
   priceEarningsToGrowthRatio: number
   priceSalesRatio: number
   dividendYield: number
   enterpriseValueMultiple: number
   priceFairValue: number
}